+++ packages/botbuilder-adapter-webex/CHANGELOG.md
@@ -1,5 +1,11 @@
 # botbuilder-adapter-webex changelog
 
+# 1.0.5
+
+* Update to latest Webex API client library.
+* Fix to mention stripping to better support Webex web client formatting
+* Update dependencies (Bot Framework to 4.6, Botkit to 4.6)
+
 # 1.0.3
 
 * Add support for [Adaptive Cards](https://developer.webex.com/blog/new-buttons-cards-webex-teams) thanks to [@darrenparkinson](https://github.com/howdyai/botkit/pull/1760)